post-processing: Move base class stage to post_processing_stages library
This is because the table containing all the stage names and their
create functions has to be defined in the same library as the code
that registers the stages, as otherwise it might not be initialised.
