+++ core/CMakeLists.txt
@@ -7,7 +7,7 @@ find_package(Boost REQUIRED COMPONENTS program_options)
 add_custom_target(VersionCpp ${CMAKE_COMMAND} -P ${CMAKE_CURRENT_LIST_DIR}/version.cmake)
 set_source_files_properties(version.cpp PROPERTIES GENERATED 1)
 
-add_library(libcamera_app libcamera_app.cpp post_processing_stage.cpp post_processor.cpp version.cpp)
+add_library(libcamera_app libcamera_app.cpp post_processor.cpp version.cpp)
 add_dependencies(libcamera_app VersionCpp)
 
 set_target_properties(libcamera_app PROPERTIES PREFIX "" IMPORT_PREFIX "")
