+++ post_processing_stages/annotate_cv_stage.cpp
@@ -11,7 +11,8 @@
 
 #include "core/frame_info.hpp"
 #include "core/libcamera_app.hpp"
-#include "core/post_processing_stage.hpp"
+
+#include "post_processing_stages/post_processing_stage.hpp"
 
 #include "opencv2/core.hpp"
 #include "opencv2/imgproc.hpp"
