+++ post_processing_stages/face_detect_cv_stage.cpp
@@ -13,9 +13,12 @@
 #include <memory>
 #include <vector>
 
-#include "../core/libcamera_app.hpp"
-#include "../core/post_processing_stage.hpp"
-#include "libcamera/geometry.h"
+#include <libcamera/geometry.h>
+
+#include "core/libcamera_app.hpp"
+
+#include "post_processing_stages/post_processing_stage.hpp"
+
 #include "opencv2/imgproc.hpp"
 #include "opencv2/objdetect.hpp"
