+++ post_processing_stages/post_processing_stage.hpp
