+++ post_processing_stages/sobel_cv_stage.cpp
@@ -7,8 +7,10 @@
 
 #include <libcamera/stream.h>
 
-#include "../core/libcamera_app.hpp"
-#include "../core/post_processing_stage.hpp"
+#include "core/libcamera_app.hpp"
+
+#include "post_processing_stages/post_processing_stage.hpp"
+
 #include "opencv2/core.hpp"
 #include "opencv2/imgproc.hpp"
